+++ RELEASE_NOTES.md
@@ -254,3 +254,4 @@ _(**NOT BACKWARDS COMPATIBLE**)_
 
 ## Version 3.2.2
 * Add support for providing an optional java.net proxy for. (#151)
+* Added support for testing serial comms on Mac/Linux